requirements.txt
Django>=3.0,<4.0
djangorestframework
drf-yasg
psycopg2-binary==2.9.10
djangorestframework-simplejwt
